backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
pydantic-settings>=2.0.0
sqlalchemy>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
python-dotenv>=1.0.0
langchain>=0.3.0
langgraph>=0.2.0
langchain-openai>=0.2.0
pinecone-client>=3.0.0
geopy>=2.4.0
duckduckgo-search>=6.0.0
